Extracting the replayer specific utilities into a separate file for readability.
